refactor: 회원정보 메서드 이전
