+++ src/main/java/com/nhnacademy/marketgg/client/dto/response/DeliveryAddressResponse.java
@@ -1,0 +1,2 @@
+package com.nhnacademy.marketgg.client.dto.response;public class DeliveryAddressResponse {
+}
